+++ embed/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 1.7.0
 
 - Regenerate the code when the content of an embedded file changes ([#53](https://github.com/fujidaiti/embed.dart/pull/53))
+- Fix embedded file paths being resolved against the workspace root directory
+  instead of the package root directory in a pub workspace ([#55])
+
+[#55]: https://github.com/fujidaiti/embed.dart/issues/55
 
 ## 1.6.7